heroku/shrimp-python-2/requirements.txt
requests
Flask==2.1.0
gunicorn==19.9.0
line-bot-sdk==1.14.0
